refactor : 컴포넌트명 ProposalHistory -> SentProposal 변경
