+++ src/pages/Proposal/SentProposal.styles.ts
