+++ src/routes/path.ts
@@ -11,5 +11,5 @@ export const RouterPath = {
   preference: '/preference',
   clientFinder: '/clientFinder',
   proposal: '/proposal/:userId',
-  proposalHistory: '/proposalHistory',
+  sentProposal: '/sentProposal',
 };
